Project1 Interactive Password Suite/requirements.txt
requests
PyYAML
zxcvbn-python
rich
